ml-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai==0.8.3
pdfplumber==0.10.3
python-docx==1.1.0
python-dotenv==1.0.0
